Fixed downloading files > 2GB

+++ src/com/turn/ttorrent/client/storage/FileCollectionStorage.java
@@ -23,8 +23,9 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-/** Multi-file torrent byte storage.
- *
+/**
+ * Multi-file torrent byte storage.
+ * 
  * <p>
  * This implementation of the torrent byte storage provides support for
  * multi-file torrents and completely abstracts the read/write operations from
@@ -32,31 +33,33 @@ import org.slf4j.LoggerFactory;
  * continuous byte storage, directly accessible by offset regardless of which
  * file this offset lands.
  * </p>
- *
+ * 
  * @author mpetazzoni
  * @author dgiffin
  */
 public class FileCollectionStorage implements TorrentByteStorage {
 
-	private static final Logger logger =
-		LoggerFactory.getLogger(FileCollectionStorage.class);
+	private static final Logger logger = LoggerFactory
+			.getLogger(FileCollectionStorage.class);
 
 	private final List<FileStorage> files;
 	private final long size;
 
-	/** Initialize a new multi-file torrent byte storage.
-	 *
-	 * @param files The list of individual {@link FileStorage}
-	 * objects making up the torrent.
-	 * @param size The total size of the torrent data, in bytes.
+	/**
+	 * Initialize a new multi-file torrent byte storage.
+	 * 
+	 * @param files
+	 *            The list of individual {@link FileStorage} objects making up
+	 *            the torrent.
+	 * @param size
+	 *            The total size of the torrent data, in bytes.
 	 */
-	public FileCollectionStorage(List<FileStorage> files,
-		long size) {
+	public FileCollectionStorage(List<FileStorage> files, long size) {
 		this.files = files;
 		this.size = size;
 
-		logger.debug("Initialized torrent byte storage on {} file(s) " +
-			"({} total byte(s)).", files.size(), size);
+		logger.debug("Initialized torrent byte storage on {} file(s) "
+				+ "({} total byte(s)).", files.size(), size);
 	}
 
 	@Override
@@ -70,8 +73,8 @@ public class FileCollectionStorage implements TorrentByteStorage {
 		int bytes = 0;
 
 		for (FileOffset fo : this.select(offset, requested)) {
-			buffer.limit(bytes + (int)fo.length);
-			bytes += fo.file.read(buffer, (int)fo.offset);
+			buffer.limit(bytes + (int) fo.length);
+			bytes += fo.file.read(buffer, fo.offset);
 		}
 
 		if (bytes < requested) {
@@ -88,8 +91,8 @@ public class FileCollectionStorage implements TorrentByteStorage {
 		int bytes = 0;
 
 		for (FileOffset fo : this.select(offset, requested)) {
-			buffer.limit(bytes + (int)fo.length);
-			bytes += fo.file.write(buffer, (int)fo.offset);
+			buffer.limit(bytes + (int) fo.length);
+			bytes += fo.file.write(buffer, (int) fo.offset);
 		}
 
 		if (bytes < requested) {
@@ -124,13 +127,14 @@ public class FileCollectionStorage implements TorrentByteStorage {
 		return true;
 	}
 
-	/** File operation details holder.
-	 *
+	/**
+	 * File operation details holder.
+	 * 
 	 * <p>
 	 * This simple inner class holds the details for a read or write operation
 	 * on one of the underlying {@link FileStorage}s.
 	 * </p>
-	 *
+	 * 
 	 * @author dgiffin
 	 * @author mpetazzoni
 	 */
@@ -147,28 +151,33 @@ public class FileCollectionStorage implements TorrentByteStorage {
 		}
 	};
 
-	/** Select the group of files impacted by an operation.
-	 *
+	/**
+	 * Select the group of files impacted by an operation.
+	 * 
 	 * <p>
 	 * This function selects which files are impacted by a read or write
 	 * operation, with their respective relative offset and chunk length.
 	 * </p>
-	 *
-	 * @param offset The offset of the operation, in bytes, relative to the
-	 * complete byte storage.
-	 * @param length The number of bytes to read or write.
-	 * @return A list of {@link FileOffset} objects representing the {@link
-	 * FileStorage}s impacted by the operation, bundled with their
-	 * respective relative offset and number of bytes to read or write.
-	 * @throws IllegalArgumentException If the offset and length go over the
-	 * byte storage size.
-	 * @throws IllegalStateException If the files registered with this byte
-	 * storage can't accomodate the request (should not happen, really).
+	 * 
+	 * @param offset
+	 *            The offset of the operation, in bytes, relative to the
+	 *            complete byte storage.
+	 * @param length
+	 *            The number of bytes to read or write.
+	 * @return A list of {@link FileOffset} objects representing the
+	 *         {@link FileStorage}s impacted by the operation, bundled with
+	 *         their respective relative offset and number of bytes to read or
+	 *         write.
+	 * @throws IllegalArgumentException
+	 *             If the offset and length go over the byte storage size.
+	 * @throws IllegalStateException
+	 *             If the files registered with this byte storage can't
+	 *             accomodate the request (should not happen, really).
 	 */
 	private List<FileOffset> select(long offset, int length) {
 		if (offset + length > this.size) {
-			throw new IllegalArgumentException("Buffer overrun (" +
-				offset + " + " + length + " > " + this.size + ") !");
+			throw new IllegalArgumentException("Buffer overrun (" + offset
+					+ " + " + length + " > " + this.size + ") !");
 		}
 
 		List<FileOffset> selected = new LinkedList<FileOffset>();
@@ -185,16 +194,15 @@ public class FileCollectionStorage implements TorrentByteStorage {
 
 			long position = offset - file.offset();
 			position = position > 0 ? position : 0;
-			int size = Math.min(
-				(int)(file.size() - position),
-				(int)(length - bytes));
+			int size = (int) Math.min((long) (file.size() - position),
+					(long) (length - bytes));
 			selected.add(new FileOffset(file, position, size));
 			bytes += size;
 		}
 
 		if (selected.size() == 0 || bytes < length) {
-			throw new IllegalStateException("Buffer underrun (only got " +
-				bytes + " out of " + length + " byte(s) requested)!");
+			throw new IllegalStateException("Buffer underrun (only got "
+					+ bytes + " out of " + length + " byte(s) requested)!");
 		}
 
 		return selected;
